chore: clean up lp artifacts and hooks
